requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.0.0
numpy>=1.18.0
matplotlib>=3.1.0
seaborn>=0.10.0
sqlparse>=0.3.0
pymysql>=0.9.3
faker>=4.0.0 